+++ CHANGELOG.rst
@@ -13,8 +13,9 @@ may want to subscribe to `GitHub's tag feed
 not released
 
 * DROPPED support for Python 3.5
-* CHANGE: ikhal: tab (and shift tab) jump from the events back to the calendar
+* CHANGE ikhal: tab (and shift tab) jump from the events back to the calendar
 * NEW Add symbol for events with at least one alarm
+* FIX URL can now be set/updated from ikhal
 
 0.10.3
 ======